moved shared nestjs files to common module
