+++ backend/src/accounts/entities/key.entity.ts
@@ -1,5 +1,5 @@
 import { Column, Entity } from "typeorm";
-import { PollingEntity } from "../../shared/entities/polling.entity";
+import { PollingEntity } from "../../common/entities/polling.entity";
 
 @Entity({ name: "keys" })
 export class AccountKey extends PollingEntity {
