+++ backend/src/common/entities/polling.entity.ts
