+++ backend/src/common/interceptors/polling-response.interceptor.ts
