+++ backend/src/events/entities/event.entity.ts
@@ -1,4 +1,4 @@
-import { PollingEntity } from "../../shared/entities/polling.entity";
+import { PollingEntity } from "../../common/entities/polling.entity";
 import { Column, Entity, PrimaryColumn } from "typeorm";
 
 @Entity({ name: "events" })
